# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: SuperNØva Smart Locker API
  description: |
    API RESTful per il sistema di smart locker SuperNØva.

    ## Autenticazione
    L'API utilizza JWT (JSON Web Tokens) per l'autenticazione. Includi il token nell'header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - 100 richieste per 15 minuti per IP
    - Risposta 429 quando il limite viene superato

    ## Errori
    Tutte le risposte di errore seguono il formato:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Descrizione dell'errore"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: SuperNØva Support
    email: api@supernova-locker.it
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.supernova-locker.it/v1
    description: Production server
...
paths:
...
  /logs:
    get:
      tags: [Logs]
      summary: Lista log accessi
      security:
        - bearerAuth: []
      parameters:
        - name: device_id
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
            enum: [unlock, lock, all]
        - name: granted
          in: query
          schema:
            type: boolean
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista log
  /logs/{deviceId}:
    get:
      tags: [Logs]
      summary: Log accessi per dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Log dispositivo
  /logs/{deviceId}/stats:
    get:
      tags: [Logs]
      summary: Statistiche accessi
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Statistiche
  /logs/export/csv:
    get:
      tags: [Logs]
      summary: Esporta log in CSV
      security:
        - bearerAuth: []
      responses:
        '200':
          description: File CSV
          content:
            text/csv:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    deviceId:
      name: deviceId
      in: path
      required: true
      description: ID univoco del dispositivo
      schema:
        type: string
        example: dev_cassapanca_001
...